LayoutTests/media/media-can-play-case-insensitive-expected.txt


Test HTMLMediaElement canPlayType() method.

EXPECTED (video.canPlayType('video/mp4') === video.canPlayType('video/MP4') == 'true') OK
EXPECTED (video.canPlayType('VIDEO/mp4') === video.canPlayType('video/mp4') == 'true') OK
EXPECTED (video.canPlayType('video/ogg') === video.canPlayType('video/OGG') == 'true') OK
EXPECTED (video.canPlayType('video/ogg') === video.canPlayType('VIDEO/ogg') == 'true') OK
EXPECTED (video.canPlayType('video/webm') === video.canPlayType('video/WEBM') == 'true') OK
EXPECTED (video.canPlayType('video/webm') === video.canPlayType('VIDEO/webm') == 'true') OK
END OF TEST
